# lib/api-zod/src/generated/types/createSaleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSaleItemBody } from "./createSaleItemBody";

export interface CreateSaleBody {
  billNo: string;
  date: string;
  customer: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  note?: string | null;
  paymentMethod: string;
  items: CreateSaleItemBody[];
}