# lib/api-zod/src/generated/types/sale.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SaleItem } from "./saleItem";

export interface Sale {
  id: number;
  billNo: string;
  date: string;
  customer: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  note: string | null;
  paymentMethod: string;
  total: number;
  createdAt: string;
  items: SaleItem[];
}