# lib/api-zod/src/generated/types/saleItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SaleItem {
  id: number;
  saleId: number;
  name: string;
  size: string;
  qty: number;
  rate: number;
  amount: number;
}